do not crash when missing coordinates
Have encountered a few games which seemed to have missing coordinates.
This seems to be a transient issue on the API-side, as subsequent
queries do not hit this snag. This change prints out the errant data and
avoids crashing, simply surfacing the remaining valid data. Unclear the
route cause, but other parties have been noticing odd features to the
NHL play-by-play data recently, so perhaps part of larger, server-side,
issue.